Deleted unnecessary files; Changed the way the cpp20 lexer is included
